refactor: :recycle: add slint rule, update imports, set Layout as cross component
