+++ components/date/date.tsx
@@ -1,10 +1,10 @@
-import { parseISO, format } from 'date-fns'
+import { parseISO, format } from 'date-fns';
 
 interface DateProps {
   dateString: string
 }
 
 export const Date = ({ dateString }: DateProps) => {
-  const date = parseISO(dateString)
-  return <time dateTime={dateString}>{format(date, 'LLLL d, yyyy')}</time>
-}
+  const date = parseISO(dateString);
+  return <time dateTime={dateString}>{format(date, 'LLLL d, yyyy')}</time>;
+};
