+++ components/date/index.ts
@@ -1,1 +1,1 @@
-export { Date } from './Date'
+export { Date } from './Date';
