+++ pages/index.tsx
@@ -1,24 +1,25 @@
-import Head from 'next/head'
-import Link from 'next/link'
-import { GetStaticProps } from 'next'
-
-import { Date } from '@/components/Date'
-import { Layout, siteTitle } from '@/components/Layout'
-import utilStyles from '../styles/utils.module.css'
-import { getSortedPostsData } from '../lib/posts'
-import STRINGS from '@/constants/string'
+import Head from 'next/head';
+import Link from 'next/link';
+import { GetStaticProps } from 'next';
+import { Date } from '@/components/Date';
+import { siteTitle } from '@/components/Layout';
+import utilStyles from '../styles/utils.module.css';
+import { getSortedPostsData } from '../utils/posts';
+import STRINGS from '@/constants/string';
 
 interface HomeProps {
   allPostsData: any
 }
+const name = 'Felipe Pardo';
 
 export default function Home({ allPostsData }: HomeProps) {
-  const postsData: { id: string; date: string; title: string }[] = allPostsData
+  const postsData: { id: string; date: string; title: string }[] = allPostsData;
   return (
-    <Layout home>
+    <>
       <Head>
         <title>{siteTitle}</title>
       </Head>
+      <h1 className={utilStyles.heading2Xl}>{name}</h1>
       <section className={utilStyles.headingMd}>
         <p className={utilStyles.justifyText}>
           {STRINGS.HOME.PROFILE_DESCRIPTION}
@@ -40,15 +41,15 @@ export default function Home({ allPostsData }: HomeProps) {
           ))}
         </ul>
       </section>
-    </Layout>
-  )
+    </>
+  );
 }
 
 export const getStaticProps: GetStaticProps = async () => {
-  const allPostsData = getSortedPostsData()
+  const allPostsData = getSortedPostsData();
   return {
     props: {
       allPostsData,
     },
-  }
-}
+  };
+};
