+++ pages/posts/[id].tsx
@@ -1,10 +1,11 @@
-import Layout from '../../components/Layout/Layout'
-import { getAllPostIds, getPostData } from '../../lib/posts'
-import Date from '../../components/Date/Date'
-import utilStyles from '../../styles/utils.module.css'
 
-import { GetStaticProps, GetStaticPaths, GetServerSideProps } from 'next'
-import Head from 'next/head'
+import { GetStaticProps, GetStaticPaths } from 'next';
+import Head from 'next/head';
+import { getAllPostIds, getPostData } from '@/utils/posts';
+import { Date } from '@/components/Date';
+import utilStyles from '@/styles/utils.module.css';
+
+
 
 export default function Post({
   postData,
@@ -16,7 +17,7 @@ export default function Post({
   }
 }) {
   return (
-    <Layout>
+    <>
       <Head>
         <title>{postData.title}</title>
       </Head>
@@ -27,23 +28,23 @@ export default function Post({
         </div>
         <div dangerouslySetInnerHTML={{ __html: postData.contentHtml }} />
       </article>
-    </Layout>
-  )
+    </>
+  );
 }
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  const paths = getAllPostIds()
+  const paths = getAllPostIds();
   return {
     paths,
     fallback: false,
-  }
-}
+  };
+};
 
 export const getStaticProps: GetStaticProps = async ({ params }) => {
-  const postData = await getPostData(params?.id as string)
+  const postData = await getPostData(params?.id as string);
   return {
     props: {
       postData,
     },
-  }
-}
+  };
+};
